fix some things

+++ Rectify11CPL/Rectify11CPL.h
@@ -36,4 +36,4 @@ extern HINSTANCE g_hInst;
 void DllAddRef();
 void DllRelease();
 
-LPCWSTR GetString(int id);
+LPCWSTR GetString(WORD id);
